Issue #501: Move JSON extensions to package android.org.json
Closes #501: Move JSON extensions to package android.org.json
