+++ components/support/ktx/src/main/java/mozilla/components/support/ktx/android/org/json/JSONObject.kt
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package mozilla.components.support.ktx.kotlin
+package mozilla.components.support.ktx.android.org.json
 
 import org.json.JSONObject
 import java.util.TreeMap
