+++ components/support/ktx/src/test/java/mozilla/components/support/ktx/android/org/json/JSONArrayTest.kt
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package mozilla.components.support.ktx.kotlin
+package mozilla.components.support.ktx.android.org.json
 
 import junit.framework.Assert.assertEquals
 import junit.framework.Assert.assertTrue
